pyproject.toml
[tool.poetry]
name = "phi3-umie"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
datasets = "^2.21.0"
umie-datasets = "^0.1.4"
click = "^8.1.7"
matplotlib = "^3.9.2"
polars = "^1.6.0"
pyter = "^0.2.2.1"
typer = "^0.12.5"
openai = "^1.43.0"
json-repair = "^0.28.4"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
